lshl/lshr value should be int

+++ jvm_rust/src/interpreter.rs
@@ -380,13 +380,13 @@ impl Interpreter {
                     stack_frame.operand_stack.push(JavaValue::Long(value1 % value2));
                 }
                 Opcode::Lshl => {
-                    let value2: i64 = stack_frame.operand_stack.pop().unwrap().into();
+                    let value2: i32 = stack_frame.operand_stack.pop().unwrap().into();
                     let value1: i64 = stack_frame.operand_stack.pop().unwrap().into();
 
                     stack_frame.operand_stack.push(JavaValue::Long(value1 << value2));
                 }
                 Opcode::Lshr => {
-                    let value2: i64 = stack_frame.operand_stack.pop().unwrap().into();
+                    let value2: i32 = stack_frame.operand_stack.pop().unwrap().into();
                     let value1: i64 = stack_frame.operand_stack.pop().unwrap().into();
 
                     stack_frame.operand_stack.push(JavaValue::Long(value1 >> value2));
